Remove dev-dependencies league/container and yiisoft/di.
